common/requirements/base.txt
alembic==1.14.0
# argon2-cffi==23.1.0
asyncpg==0.30.0
# bcrypt==4.2.1
# disposable-email-domains==0.0.117
# fastapi==0.115.6
# fastapi-mail==1.4.2
# httpx==0.28.1
# PyJWT==2.10.1
loguru==0.7.3
# psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic-settings==2.7.1
pydantic[email]
python-decouple==3.8
# redis==5.2.1
sqlmodel[asyncio]==0.0.22
# uvicorn==0.34.0
setuptools==80.9.0